Flasks/requirements.txt
<!--Moved to Frontend and Backend Folder after microservices topic-->

flask
pymongo
dnspython
python-dotenv